+++ modules/client/src/multisig/internal/graphql-queries/proposal.ts
@@ -18,12 +18,17 @@ query MultisigProposal($proposalId: ID!) {
       value
       data
     }
+    executionDate
+    executionBlockNumber
+    creationBlockNumber
+    plugin {
+      minApprovals
+      onlyListed
+    }
     executionTxHash
     executed
     approvers{
-      approver{
-        id
-      }
+      id
     }
   }
 }
@@ -42,6 +47,16 @@ query MultisigProposals($where: MultisigProposal_filter!, $limit:Int!, $skip: In
     approvals
     startDate
     endDate
+    executionDate
+    executionBlockNumber
+    creationBlockNumber
+    approvers {
+      id
+    }
+    plugin{
+      minApprovals
+      onlyListed
+    }
   }
 }
 `;